Support running the desired program with a launcher
* flo_check_homework/main.py: support a new setting called
ProgramLauncher in the configuration file which, if non-empty, specifies
a launcher program. For instance, if ProgramLauncher=/path/to/foo, then
instead of running 'desired-prog arg1 ... argn', flo-check-homework will
run '/path/to/foo desired-prog arg1 ... argn'. If ProgramLauncher is
empty, then no launcher is used, as was the case before this commit.