Code reorganization to prepare for the "super magic word" feature
* flo_check_homework/main.py: move code related to the desired program
(initialization, starting, checking the exit status) from the MainWindow
class to the HomeWorkCheckApp class. This will be useful to the next
commit, but requires other things to move, most notably the
initialization of the QSettings, which is needed to read the
ProgramLauncher setting, but should be done after the lock file has been
acquired in order to avoid concurent access to the QSettings. This is
why HomeWorkCheckApp now has an endInit() method that should only be
called after the lock file has been acquired (and also after command
line parsing because of the need to know the name of the desired
program).

* flo_check_homework/main.py: move command line parsing from a
module-level function to HomeWorkCheckApp.processCommandLine().

* flo_check_homework/main.py: move the calls to
app.setOrganizationName(), app.setOrganizationDomain(),
app.setApplicationName() and the application icon setup from
module-level to HomeWorkCheckApp.__init__().