New algorithm to determine the duration of the super magic token
* flo_check_homework/main.py: when the super magic word entered by the
user contains a decimal point, the duration of the super magic token is
determined using a new method. This should make it practically
impossible for unauthorized personnel to tamper with the token duration
as the super magic word is being entered.

* flo_check_homework/main.py: use the decimal module to ensure that the
digits seen by the program are the same as those entered by the user.